Reorganize files and add helper version
